@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net7.0;net6.0</TargetFrameworks>
-    <LangVersion>11.0</LangVersion>
-    <Version>7.0.0</Version>
+    <TargetFrameworks>net8.0;net7.0;net6.0</TargetFrameworks>
+    <Version>8.0.0</Version>
+    <IsAotCompatible>true</IsAotCompatible>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\NCoreUtils.Extensions.Memory\NCoreUtils.Extensions.Memory.csproj" />
